fastapi-app/requirements.txt
fastapi
uvicorn
requests
httpx
python-jose[cryptography]   # ✅ adds jose
redis
django-environ
